Use new build-code package to implement agent-kit substitute component
