+++ packages/agent-kit/.gitignore
@@ -1,1 +1,2 @@
 graphs
+src/generated
