+++ packages/agent-kit/src/boards/specialist.ts
@@ -6,7 +6,6 @@
 
 import {
   annotate,
-  anyOf,
   array,
   board,
   enumeration,
@@ -16,7 +15,6 @@ import {
   output,
   outputNode,
   starInputs,
-  string,
   Value,
 } from "@breadboard-ai/build";
 import { code, coreKit } from "@google-labs/core-kit";
@@ -39,7 +37,6 @@ import {
   boardInvocationArgsType,
   boardInvocationAssemblerFunction,
   functionDeclarationsFormatterFn,
-  functionDeclarationType,
   functionOrTextRouterFunction,
   type FunctionSignatureItem,
   responseCollatorFunction,
@@ -50,7 +47,7 @@ import boardToFunction from "./internal/board-to-function.js";
 import invokeBoardWithArgs from "./internal/invoke-board-with-args.js";
 import specialistDescriber from "./internal/specialist-describer.js";
 import { GenericBoardDefinition } from "@breadboard-ai/build/internal/board/board.js";
-import { substitute } from "../templating.js";
+import { substitute } from "../generated/substitute.js";
 
 const inputs = starInputs({ type: object({}, "unknown") });
 
@@ -85,22 +82,13 @@ const model = input({
   examples: ["gemini-1.5-flash-latest"],
 });
 
-const substituteParams = code(
-  {
-    $metadata: {
-      title: "Substitute Parameters",
-      description: "Performing parameter substitution, if needed.",
-    },
-    "*": inputs,
+const substituteParams = substitute({
+  $metadata: {
+    title: "Substitute Parameters",
+    description: "Performing parameter substitution, if needed.",
   },
-  {
-    in: array(contextType),
-    persona: anyOf(llmContentType, string({})),
-    task: anyOf(llmContentType, string({})),
-    outs: array(functionDeclarationType),
-  },
-  substitute
-);
+  "*": inputs,
+});
 
 const addTask = code(
   {
